convDL benchmarked
