+++ src/models/deepseekr1_model.py
@@ -9,9 +9,6 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
-from langchain.prompts import PromptTemplate
-from langchain.schema.runnable import Runnable
-from peft import PromptTuningConfig, PromptTuningInit, TaskType, get_peft_model
 from transformers import (AutoModelForCausalLM, AutoTokenizer,
                           BitsAndBytesConfig)
